+++ src/Contracts/Priceable.php
@@ -4,7 +4,7 @@ namespace Jegex\LaravelPriceable\Contracts;
 
 use Illuminate\Database\Eloquent\Relations\MorphMany;
 
-interface Purchasable
+interface Priceable
 {
     public function prices(): MorphMany;