+++ tests/Models/Product.php
@@ -3,10 +3,10 @@
 namespace Jegex\LaravelPriceable\Tests\Models;
 
 use Illuminate\Database\Eloquent\Model;
-use Jegex\LaravelPriceable\Contracts\Purchasable;
+use Jegex\LaravelPriceable\Contracts\Priceable;
 use Jegex\LaravelPriceable\Traits\HasPrices;
 
-class Product extends Model implements Purchasable
+class Product extends Model implements Priceable
 {
     use HasPrices;